주문 시 발생하는 오류 수정

+++ src/main/java/com/example/flora_shop/controller/ItemForm.java
@@ -6,6 +6,6 @@ import lombok.Setter;
 @Getter
 @Setter
 public class ItemForm {
-    private int id;
-    private int quantity;
+    private Long id;
+    private Long quantity;
 }
